+++ src/main/java/com/hanaieum/server/domain/bucketList/dto/BucketListDetailResponse.java
@@ -26,7 +26,7 @@ public class BucketListDetailResponse {
                     .accountId(bucketList.getMoneyBoxAccount().getId())
                     .boxName(bucketList.getMoneyBoxAccount().getBoxName())
                     .accountNumber(bucketList.getMoneyBoxAccount().getNumber())
-                    .balance(BigDecimal.valueOf(bucketList.getMoneyBoxAccount().getBalance()))
+                    .balance(bucketList.getMoneyBoxAccount().getBalance())
                     .hasMoneyBox(true)
                     .build();
         } else {
